final commit

+++ server/src/index.ts
@@ -6,6 +6,12 @@ import { expressMiddleware } from "@as-integrations/express5";
 import { connectDB } from "./db/connection";
 import { Project } from "./model/Project";
 import { Task } from "./model/Task";
+import bcrypt from "bcryptjs";
+import jwt from "jsonwebtoken";
+import { User } from "./model/User";
+import { register } from "module";
+import { getUserIdFromToken } from "./utils/auth";
+import cors from "cors";
 
 dotenv.config({
   path: "../.env",
@@ -13,6 +19,13 @@ dotenv.config({
 
 dns.setServers(["8.8.8.8", "8.8.4.4"]);
 const app = express();
+app.use(
+  cors({
+    origin: "http://localhost:5173",
+  }),
+);
+
+
 const PORT = process.env.PORT;
 
 const server = new ApolloServer({
@@ -27,6 +40,16 @@ const server = new ApolloServer({
     id: ID!
     title: String!
     completed: Boolean!
+    }
+    type User {
+    id: ID!
+    name: String!
+    email: String!
+    }
+
+    type AuthPayload {
+    user: User!
+    token: String!
     }
 
     input CreateProjectInput{
@@ -48,6 +71,17 @@ const server = new ApolloServer({
       title: String,
       completed: Boolean
     }
+    
+    input RegisterInput {
+    name: String!
+    email: String!
+    password: String!
+    }
+
+    input LoginInput {
+    email: String!
+    password: String!
+    }
     type Query{
         hello: String
         goodbye: String
@@ -65,12 +99,20 @@ const server = new ApolloServer({
         createTask(input: CreateTaskInput!): Task!
         updateTask(id: ID!, input: UpdateTaskInput!): Task!
         deleteTask(id: ID!): Task!
+        register(input: RegisterInput!): AuthPayload!
+        login(input: LoginInput!): AuthPayload! 
     }
     
     `,
   resolvers: {
     Query: {
-      hello: () => "Hello from GraphQL!",
+      hello: (_: unknown, __: unknown, context: { userId: string | null }) => {
+        if (!context.userId) {
+          return "You are not logged in";
+        }
+
+        return `You are user ${context.userId}`;
+      },
       goodbye: () => "Goodbye from GraphQL!",
       message: () => "This is my first GraphQL project",
       projects: async () => await Project.find(),
@@ -93,7 +135,11 @@ const server = new ApolloServer({
             description: string;
           };
         },
+        context: { userId: string | null },
       ) => {
+        if (!context.userId) {
+          throw new Error("You must be logged in");
+        }
         const project = await Project.create({
           name: args.input.name,
           description: args.input.description,
@@ -109,7 +155,12 @@ const server = new ApolloServer({
             description?: string;
           };
         },
+        context: { userId: string | null },
       ) => {
+        if (!context.userId) {
+          throw new Error("You must be logged in");
+        }
+
         const project = await Project.findByIdAndUpdate(
           args.id,
           {
@@ -118,9 +169,18 @@ const server = new ApolloServer({
           },
           { new: true },
         );
+
         return project;
       },
-      deleteProject: async (_: unknown, args: { id: string }) => {
+      deleteProject: async (
+        _: unknown,
+        args: { id: string },
+        context: { userId: string | null },
+      ) => {
+        if (!context.userId) {
+          throw new Error("You must be logged in");
+        }
+
         const project = await Project.findByIdAndDelete(args.id);
 
         if (project) {
@@ -139,11 +199,17 @@ const server = new ApolloServer({
             title: string;
           };
         },
+        context: { userId: string | null },
       ) => {
+        if (!context.userId) {
+          throw new Error("You must be logged in");
+        }
+
         const task = await Task.create({
           projectId: args.input.projectId,
           title: args.input.title,
         });
+
         return task;
       },
       updateTask: async (
@@ -155,7 +221,12 @@ const server = new ApolloServer({
             completed: boolean;
           };
         },
+        context: { userId: string | null },
       ) => {
+        if (!context.userId) {
+          throw new Error("You must be logged in");
+        }
+
         const task = await Task.findByIdAndUpdate(
           args.id,
           {
@@ -164,13 +235,92 @@ const server = new ApolloServer({
           },
           { new: true },
         );
+
         return task;
       },
-      deleteTask: async(_:unknown,args:{id:string})=>{
-        const task=await Task.findByIdAndDelete(args.id)
-        return task
-      }
+      deleteTask: async (
+        _: unknown,
+        args: { id: string },
+        context: { userId: string | null },
+      ) => {
+        if (!context.userId) {
+          throw new Error("You must be logged in");
+        }
+
+        const task = await Task.findByIdAndDelete(args.id);
+
+        return task;
+      },
+
+      register: async (
+        _: unknown,
+        args: {
+          input: {
+            name: string;
+            email: string;
+            password: string;
+          };
+        },
+      ) => {
+        const { name, email, password } = args.input;
+        const existingUser = await User.findOne({ email });
+        if (existingUser) {
+          throw new Error("User already exists");
+        }
+        const hashedPassword = await bcrypt.hash(password, 10);
+
+        const user = User.create({
+          name,
+          email,
+          password: hashedPassword,
+        });
+
+        const token = jwt.sign(
+          { userId: (await user)._id.toString() },
+          process.env.JWT_SECRET!,
+          { expiresIn: "7d" },
+        );
+        return {
+          user,
+          token,
+        };
+      },
+      login: async (
+        _: unknown,
+        args: {
+          input: {
+            email: string;
+            password: string;
+          };
+        },
+      ) => {
+        const { email, password } = args.input;
+
+        const user = await User.findOne({ email });
+
+        if (!user) {
+          throw new Error("Invalid email or password");
+        }
+
+        const passwordMatch = await bcrypt.compare(password, user.password);
+
+        if (!passwordMatch) {
+          throw new Error("Invalid email or password");
+        }
+
+        const token = jwt.sign(
+          { userId: user._id.toString() },
+          process.env.JWT_SECRET!,
+          { expiresIn: "7d" },
+        );
+
+        return {
+          user,
+          token,
+        };
+      },
     },
+
     Project: {
       id: (project: any) => project._id.toString(),
       tasks: async (project: any) => {
@@ -189,7 +339,33 @@ async function startServer() {
   await connectDB();
   await server.start();
 
-  app.use("/graphql", express.json(), expressMiddleware(server));
+  app.use(
+    "/graphql",
+    express.json(),
+    expressMiddleware(server, {
+      context: async ({ req }) => {
+
+        const authHeader = req.headers.authorization;
+
+        if (!authHeader) {
+          console.log("NO AUTH HEADER");
+          return { userId: null };
+        }
+
+        const token = authHeader.split(" ")[1];
+
+        try {
+          const userId = getUserIdFromToken(token);
+
+          return { userId };
+        } catch (error) {
+          console.log("TOKEN VERIFICATION FAILED:", error);
+
+          return { userId: null };
+        }
+      },
+    }),
+  );
 
   app.listen(PORT, () => {
     console.log(`Server started listening on port ${PORT}`);
